updated style-guide, refactored accordingly
