+++ src/components/PayPal/README.md
