+++ src/pages/_app.tsx
@@ -1,9 +1,9 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 
-import NavBar from "@/components/NavBar/NavBar";
+import NavBar from "@/components/NavBar";
 
-import PanicButton from "@/components/panicbutton/panicbutton";
+import PanicButton from "@/components/PanicButton";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
