+++ src/pages/core-services.tsx
@@ -1,12 +1,12 @@
 import Link from "next/link";
 
-import InPageNav from "@/components/inpagenav";
+import InPageNav from "@/components/InPageNav";
 
-import CounsellingSection from "../components/services/core-services/counselling";
-import PoliceSection from "../components/services/core-services/police-stations";
-import RefugesSection from "../components/services/core-services/refuges";
-import SafeAtHomeSection from "../components/services/core-services/safe-at-home";
-import TransitionalHousingSection from "../components/services/core-services/transitional-housing";
+import CounsellingSection from "../components/Services/CoreServices/counselling";
+import PoliceSection from "../components/Services/CoreServices/police-stations";
+import RefugesSection from "../components/Services/CoreServices/refuges";
+import SafeAtHomeSection from "../components/Services/CoreServices/safe-at-home";
+import TransitionalHousingSection from "../components/Services/CoreServices/transitional-housing";
 
 export default function CoreServices() {
   const navLinkDetails = [
